remove all spot cam ros components and move wrapper components to the python package
