--- spot_cam/launch/spot_cam.launch
@@ -1,18 +1,0 @@
-<launch>
-    <arg name="username" default="dummyusername"/>
-    <arg name="password" default="dummypassword"/>
-    <arg name="hostname" default="192.168.50.3"/>
-
-    <node name="spot_cam" pkg="spot_cam" type="spot_cam" output="screen">
-        <param name="username" value="$(arg username)"/>
-        <param name="password" value="$(arg password)"/>
-        <param name="hostname" value="$(arg hostname)"/>
-    </node>
-
-    <!-- Add a compressed version of the webrtc image stream  -->
-    <node name="compress_spot_webrtc" pkg="image_transport" type="republish" args="raw compressed">
-        <remap from="in" to="/spot/cam/image"/>
-        <remap from="out" to="/spot/cam/image"/>
-    </node>
-
-</launch>
