--- spot_cam/msg/BITStatus.msg
@@ -1,3 +1,0 @@
-# https://dev.bostondynamics.com/protos/bosdyn/api/proto_reference#getbitstatusresponse
-spot_msgs/SystemFault[] events
-spot_cam/Degradation[] degradations
