--- spot_cam/msg/PTZDescription.msg
@@ -1,9 +1,0 @@
-# https://dev.bostondynamics.com/protos/bosdyn/api/proto_reference#ptzdescription
-# Name of this ptz (can be virtual)
-string name
-# Limits in degrees on the pan axis
-spot_cam/PTZLimits pan_limit
-# Limits in degrees on the pan axis
-spot_cam/PTZLimits tilt_limit
-# Limits in degrees on the pan axis
-spot_cam/PTZLimits zoom_limit
