--- spot_cam/msg/PTZDescriptionArray.msg
@@ -1,1 +1,0 @@
-spot_cam/PTZDescription[] ptzs
