--- spot_cam/msg/PTZLimits.msg
@@ -1,8 +1,0 @@
-# https://dev.bostondynamics.com/protos/bosdyn/api/proto_reference#ptzdescription-limits
-# If both max and min are zero, this means the limit is unset. The documentation states that if a limit
-# is unset, then all positions are valid.
-# https://dev.bostondynamics.com/protos/bosdyn/api/proto_reference#ptzdescription
-# Minimum value for the axis
-float32 min
-# Maximum value for the axis
-float32 max
