--- spot_cam/msg/PTZStateArray.msg
@@ -1,1 +1,0 @@
-spot_cam/PTZState[] ptzs
