--- spot_cam/msg/StringMultiArray.msg
@@ -1,1 +1,0 @@
-string[] data
