--- spot_cam/msg/Temperature.msg
@@ -1,2 +1,0 @@
-string channel_name
-float32 temperature
