--- spot_cam/msg/TemperatureArray.msg
@@ -1,1 +1,0 @@
-spot_cam/Temperature[] temperatures
