--- spot_cam/package.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-    <name>spot_cam</name>
-    <version>0.0.0</version>
-    <description>The spot_cam package</description>
-
-    <maintainer email="michal@robots.ox.ac.uk">Michal Staniaszek</maintainer>
-
-    <license>TODO</license>
-    <author>Michal Staniaszek</author>
-
-    <buildtool_depend>catkin</buildtool_depend>
-    <depend>rospy</depend>
-    <depend>std_msgs</depend>
-    <depend>geometry_msgs</depend>
-    <depend>spot_msgs</depend>
-    <depend>spot_driver</depend>
-    <build_depend>message_generation</build_depend>
-    <exec_depend>message_runtime</exec_depend>
-</package>
