--- spot_cam/scripts/spot_cam
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-import rospy
-from spot_cam.spot_cam_ros import SpotCamROS
-
-if __name__ == '__main__':
-    rospy.init_node("spot_cam")
-    spot_cam = SpotCamROS()
-    spot_cam.main()
