--- spot_cam/srv/SetBool.srv
@@ -1,4 +1,0 @@
-bool value
-----
-bool success
-string message
