--- spot_cam/srv/SetFloat.srv
@@ -1,4 +1,0 @@
-float32 value
-----
-bool success
-string message
