--- spot_cam/srv/SetStreamParams.srv
@@ -1,4 +1,0 @@
-spot_cam/StreamParams params
-----
-bool success
-string message
