--- spot_cam/srv/SetString.srv
@@ -1,4 +1,0 @@
-string value
-----
-bool success
-string message
